tweaks for the word-level functions
