+++ spoonerize/__init__.py
@@ -3,4 +3,5 @@
 from .word import spoonerize_word_pair
 
 
+__all__ = ['regexes', 'word']
 __version__ = '0.1'
